requirements.txt
absl-py==2.1.0
accelerate==1.6.0
aiofiles==23.2.1
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
altair==5.5.0
altgraph @ file:///System/Volumes/Data/SWE/Apps/DT/BuildRoots/BuildRoot2/ActiveBuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-124/altgraph-0.17.2-py2.py3-none-any.whl
annotated-types==0.7.0
anyio==4.6.2.post1
appnope==0.1.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
async-timeout==4.0.3
attrs==24.2.0
babel==2.16.0
backoff==2.2.1
bcrypt==4.3.0
beautifulsoup4==4.12.3
bidict==0.23.1
bleach==6.1.0
blinker==1.9.0
build==1.2.2.post1
cachetools==5.5.2
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.0
chromadb==1.0.12
click==8.1.8
cohere==5.15.0
coloredlogs==15.0.1
comm==0.2.2
contourpy==1.3.0
ctransformers==0.2.27
cycler==0.12.1
dataclasses-json==0.6.7
datasets==3.0.1
debugpy==1.8.7
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.18
dill==0.3.8
distro==1.9.0
docutils==0.21.2
durationpy==0.10
exceptiongroup==1.2.2
executing==2.1.0
faiss-cpu==1.10.0
fastapi==0.115.9
fastavro==1.11.1
fastjsonschema==2.20.0
ffmpy==0.5.0
filelock==3.16.1
filetype==1.2.0
Flask==3.1.0
flatbuffers==24.3.25
fonttools==4.54.1
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2024.6.1
future @ file:///System/Volumes/Data/SWE/Apps/DT/BuildRoots/BuildRoot2/ActiveBuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-124/future-0.18.2-py3-none-any.whl
fuzzywuzzy==0.18.0
gast==0.6.0
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.174.0
google-auth==2.40.2
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
google-pasta==0.2.0
googleapis-common-protos==1.70.0
gradio==4.44.1
gradio_client==1.3.0
greenlet==3.2.2
grpcio==1.73.1
grpcio-status==1.71.0
h11==0.16.0
h5py==3.12.1
httpcore==1.0.9
httplib2==0.22.0
httptools==0.6.4
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.25.2
humanfriendly==10.0
idna==3.10
ifaddr==0.2.0
importlib_metadata==8.5.0
importlib_resources==6.4.5
ipykernel==6.29.5
ipython==8.18.1
ipywidgets==8.1.5
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.6
jiter==0.9.0
joblib==1.4.2
json5==0.9.25
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
jupyterthemes==0.20.0
kaggle==1.7.4.2
keras==3.6.0
kiwisolver==1.4.7
kubernetes==32.0.1
labelImg==1.8.6
langchain==0.3.25
langchain-cohere==0.4.4
langchain-community==0.3.24
langchain-core==0.3.63
langchain-openai==0.3.18
langchain-text-splitters==0.3.8
langsmith==0.3.43
lesscpy==0.15.1
Levenshtein==0.27.1
libclang==18.1.1
lxml==6.0.0
macholib @ file:///System/Volumes/Data/SWE/Apps/DT/BuildRoots/BuildRoot2/ActiveBuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-124/macholib-1.15.2-py2.py3-none-any.whl
Markdown==3.7
markdown-it-py==3.0.0
markdown2==2.5.3
MarkupSafe==2.1.5
marshmallow==3.26.1
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
ml-dtypes==0.3.2
mmh3==5.1.0
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
mypy_extensions==1.1.0
namex==0.0.8
narwhals==1.33.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.2.1
nicegui==2.21.0
notebook==7.2.2
notebook_shim==0.2.4
numpy==1.26.4
oauthlib==3.2.2
olefile==0.47
onnxruntime==1.19.2
openai==1.93.1
opentelemetry-api==1.33.1
opentelemetry-exporter-otlp-proto-common==1.33.1
opentelemetry-exporter-otlp-proto-grpc==1.33.1
opentelemetry-instrumentation==0.54b1
opentelemetry-instrumentation-asgi==0.54b1
opentelemetry-instrumentation-fastapi==0.54b1
opentelemetry-proto==1.33.1
opentelemetry-sdk==1.33.1
opentelemetry-semantic-conventions==0.54b1
opentelemetry-util-http==0.54b1
opt_einsum==3.4.0
optree==0.13.0
orjson==3.10.16
outcome==1.3.0.post0
overrides==7.7.0
packaging==24.1
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pi==0.1.2
pillow==10.4.0
platformdirs==4.3.6
playwright==1.53.0
ply==3.11
posthog==4.2.0
prometheus_client==0.21.0
prompt_toolkit==3.0.48
propcache==0.2.0
proto-plus==1.26.1
protobuf==5.29.5
pscript==0.7.7
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.1
pydantic-settings==2.9.1
pydantic_core==2.33.0
pydeck==0.9.1
pydub==0.25.1
pyee==13.0.0
Pygments==2.18.0
pyparsing==3.2.0
pypdf==5.6.0
PyPika==0.48.9
pyproject_hooks==1.2.0
PyQt5==5.15.11
PyQt5-Qt5==5.15.17
PyQt5_sip==12.17.0
PySocks==1.7.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-engineio==4.12.2
python-iso639==2025.2.18
python-json-logger==2.0.7
python-Levenshtein==0.27.1
python-magic==0.4.27
python-multipart==0.0.20
python-slugify==8.0.4
python-socketio==5.13.0
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
RapidFuzz==3.13.0
referencing==0.35.1
regex==2024.9.11
replicate==1.0.7
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.2
rpds-py==0.20.0
rsa==4.9.1
ruff==0.11.4
safetensors==0.4.5
scikit-learn==1.6.1
scipy==1.13.1
selenium==4.34.0
semantic-version==2.10.0
Send2Trash==1.8.3
sentence-transformers==4.1.0
sentencepiece==0.2.0
shellingham==1.5.4
simple-websocket==1.1.0
six @ file:///System/Volumes/Data/SWE/Apps/DT/BuildRoots/BuildRoot2/ActiveBuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-124/six-1.15.0-py2.py3-none-any.whl
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==2.0.41
stack-data==0.6.3
starlette==0.45.3
streamlit==1.44.1
sympy==1.13.3
tenacity==9.1.2
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.2
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.5.0
terminado==0.18.1
text-unidecode==1.3
tf_keras==2.16.0
threadpoolctl==3.5.0
tiktoken==0.9.0
tinycss2==1.3.0
tokenizers==0.20.1
toml==0.10.2
tomli==2.0.2
tomlkit==0.12.0
torch==2.2.2
torchvision==0.17.2
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
transformers==4.45.2
trio==0.30.0
trio-websocket==0.12.2
typer==0.15.2
types-python-dateutil==2.9.0.20241003
types-PyYAML==6.0.12.20250516
types-requests==2.32.0.20250602
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.14.0
tzdata==2024.2
ultralytics==8.3.159
uri-template==1.3.0
uritemplate==4.2.0
urllib3==2.4.0
uvicorn==0.34.0
uvloop==0.21.0
vbuild==0.8.2
watchdog==6.0.0
watchfiles==1.0.5
wcwidth==0.2.13
webcolors==24.8.0
webdriver-manager==4.0.2
webencodings==0.5.1
websocket-client==1.8.0
websockets==12.0
Werkzeug==3.1.3
widgetsnbextension==4.0.13
wrapt==1.16.0
wsproto==1.2.0
xxhash==3.5.0
yarl==1.15.3
zipp==3.20.2
zstandard==0.23.0
